Refactors code
