+++ apps/judicial-system/backend/src/app/modules/notification/services/defendantNotification/defendantNotification.service.ts
@@ -34,9 +34,9 @@ import {
   Notification,
   Recipient,
 } from '../../../repository'
-import { BaseNotificationService } from '../../baseNotification.service'
 import { DeliverResponse } from '../../models/deliver.response'
 import { notificationModuleConfig } from '../../notification.config'
+import { BaseNotificationService } from '../baseNotification.service'
 import { strings } from './defendantNotification.strings'
 
 @Injectable()
