+++ apps/judicial-system/backend/src/app/modules/notification/services/subpoenaNotification/subpoenaNotification.service.ts
@@ -21,9 +21,9 @@ import {
 import { CourtService } from '../../../court'
 import { EventService } from '../../../event'
 import { Case, Notification, Recipient } from '../../../repository'
-import { BaseNotificationService } from '../../baseNotification.service'
 import { DeliverResponse } from '../../models/deliver.response'
 import { notificationModuleConfig } from '../../notification.config'
+import { BaseNotificationService } from '../baseNotification.service'
 import { strings } from './subpoenaNotification.strings'
 
 @Injectable()
