+++ apps/judicial-system/backend/src/app/modules/notification/test/createTestingNotificationModule.ts
@@ -36,13 +36,13 @@ import { UserService } from '../../user'
 import { InternalNotificationController } from '../internalNotification.controller'
 import { notificationModuleConfig } from '../notification.config'
 import { NotificationController } from '../notification.controller'
-import { NotificationService } from '../notification.service'
-import { NotificationDispatchService } from '../notificationDispatch.service'
 import { CaseNotificationService } from '../services/caseNotification/caseNotification.service'
 import { CivilClaimantNotificationService } from '../services/civilClaimantNotification/civilClaimantNotification.service'
 import { DefendantNotificationService } from '../services/defendantNotification/defendantNotification.service'
 import { IndictmentCaseNotificationService } from '../services/indictmentCaseNotification/indictmentCaseNotification.service'
 import { InstitutionNotificationService } from '../services/institutionNotification/institutionNotification.service'
+import { NotificationService } from '../services/notification.service'
+import { NotificationDispatchService } from '../services/notificationDispatch.service'
 
 jest.mock('@island.is/judicial-system/message')
